Binary Exchange Maintenancde Script
Originally posted in the MS Technet Gallery but was lost in the stupid decision to shut everything down and leave it to the original authors to repost them on the new place.
We understand the concerns due to intellectual property and other things, but really?
This is NOT my work. It is from Vijaya Reddy, whom you can find here: https://www.blogger.com/profile/08822887414786847861
Since he has not posted for a long time and I have the files for my own use, I decided to post them here.